requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.2.0
matplotlib>=3.8.0
scikit-image>=0.22.0
pandas>=2.2.0
scipy>=1.12.0
fpdf2>=2.7.9
pytest>=8.1.0
